+++ demuxer.py
@@ -4,7 +4,7 @@ https://github.com/sam210723/xrit-rx
 """
 
 import ccsds as CCSDS
-from collections import deque
+from collections import deque, namedtuple
 from time import sleep
 from threading import Thread
 import sys
@@ -76,8 +76,8 @@ class Demuxer:
                 # Check for VCID change
                 if lastVCID != vcdu.VCID:
                     # Notify channel handlers of VCID change
-                    for chan in self.channels:
-                        self.channels[chan].notify(vcdu.VCID)
+                    for c in self.channels:
+                        self.channels[c].notify(vcdu.VCID)
                     
                     # Print VCID info
                     if self.config.verbose: print()
@@ -90,13 +90,14 @@ class Demuxer:
                 
                 # Discard VCDUs in blacklisted VCIDs
                 if vcdu.VCID in self.config.blacklist: continue
-                
+
                 # Check channel handler for current VCID exists
                 try:
                     self.channels[vcdu.VCID]
                 except KeyError:
                     # Create new channel handler instance
-                    self.channels[vcdu.VCID] = Channel(vcdu.VCID, self.config.verbose, crclut, self.config.output, self.config.images, self.config.xrit, self.config.keys)
+                    ccfg = namedtuple('ccfg', 'VCID verbose lut output images xrit keys')
+                    self.channels[vcdu.VCID] = Channel(ccfg(vcdu.VCID, self.config.verbose, crclut, self.config.output, self.config.images, self.config.xrit, self.config.keys))
                     if self.config.verbose: print("  CREATED NEW CHANNEL HANDLER\n")
 
                 # Pass VCDU to appropriate channel handler
@@ -151,24 +152,13 @@ class Channel:
     Virtual channel data handler
     """
 
-    def __init__(self, vcid, v, crclut, output, i, x, k):
+    def __init__(self, config):
         """
         Initialises virtual channel data handler
-        :param vcid: Virtual Channel ID
-        :param v: Verbose output flag
-        :param crclut: CP_PDU CRC LUT
-        :param output: xRIT file output path root
-        :param k: Decryption keys
         """
 
-        self.VCID = vcid            # VCID for this handler
+        self.config = config        # Configuration tuple
         self.counter = -1           # VCDU continuity counter
-        self.verbose = v            # Verbose output flag
-        self.crclut = crclut        # CP_PDU CRC LUT
-        self.outputPath = output    # xRIT file output path root
-        self.outputImages = i       # Flag for saving Images to disk
-        self.outputXRIT = x         # Flag for saving xRIT files to disk
-        self.keys = k               # Decryption keys
         self.cCPPDU = None          # Current CP_PDU object
         self.cTPFile = None         # Current TP_File object
 
@@ -193,13 +183,13 @@ class Channel:
                 preptr = mpdu.PACKET[:mpdu.POINTER]
 
                 try:
-                    lenok, crcok = self.cCPPDU.finish(preptr, self.crclut)
-                    if self.verbose: self.check_CPPDU(lenok, crcok)
+                    lenok, crcok = self.cCPPDU.finish(preptr, self.config.lut)
+                    if self.config.verbose: self.check_CPPDU(lenok, crcok)
 
                     # Handle finished CP_PDU
                     self.handle_CPPDU(self.cCPPDU)
                 except AttributeError:
-                    if self.verbose: print("  NO CP_PDU TO FINISH (DROPPED PACKETS?)")
+                    if self.config.verbose: print("  NO CP_PDU TO FINISH (DROPPED PACKETS?)")
 
                 #TODO: Check CP_PDU continuity
 
@@ -217,13 +207,13 @@ class Channel:
                     self.cCPPDU.PAYLOAD = self.cCPPDU.PAYLOAD[:self.cCPPDU.LENGTH]
                     
                     try:
-                        lenok, crcok = self.cCPPDU.finish(b'', self.crclut)
-                        if self.verbose: self.check_CPPDU(lenok, crcok)
+                        lenok, crcok = self.cCPPDU.finish(b'', self.config.lut)
+                        if self.config.verbose: self.check_CPPDU(lenok, crcok)
 
                         # Handle finished CP_PDU
                         self.handle_CPPDU(self.cCPPDU)
                     except AttributeError:
-                        if self.verbose: print("  NO CP_PDU TO FINISH (DROPPED PACKETS?)")
+                        if self.config.verbose: print("  NO CP_PDU TO FINISH (DROPPED PACKETS?)")
 
             else:
                 # First CP_PDU in TP_File
@@ -233,9 +223,9 @@ class Channel:
             # Handle special EOF CP_PDU
             if self.cCPPDU.is_EOF():
                 self.cCPPDU = None
-                if self.verbose: print("  [CP_PDU] EOF MARKER")
+                if self.config.verbose: print("  [CP_PDU] EOF MARKER")
             else:
-                if self.verbose:
+                if self.config.verbose:
                     self.cCPPDU.print_info()
                     print("    HEADER:     0x{}".format(self.cCPPDU.header.hex().upper()))
                     print("    OFFSET:     0x{}\n    ".format(hex(mpdu.POINTER)[2:].upper()), end="")
@@ -244,15 +234,15 @@ class Channel:
             try:
                 wasparsed = self.cCPPDU.PARSED
                 self.cCPPDU.append(mpdu.PACKET)
-                if wasparsed != self.cCPPDU.PARSED and self.verbose:
+                if wasparsed != self.cCPPDU.PARSED and self.config.verbose:
                     self.cCPPDU.print_info()
                     print("    HEADER:     0x{}".format(self.cCPPDU.header.hex().upper()))
                     print("    OFFSET:     SPANS MULTIPLE M_PDUs\n", end="")
             except AttributeError:
-                if self.verbose: print("  NO CP_PDU TO APPEND M_PDU TO (DROPPED PACKETS?)")
+                if self.config.verbose: print("  NO CP_PDU TO APPEND M_PDU TO (DROPPED PACKETS?)")
         
         # VCDU indicator
-        if self.verbose: print(".", end="")
+        if self.config.verbose: print(".", end="")
         sys.stdout.flush()
     
 
@@ -270,7 +260,7 @@ class Channel:
             
             diff = vcdu.COUNTER - self.counter - 1
             if diff > 0:
-                if self.verbose:
+                if self.config.verbose:
                     print("  DROPPED {} PACKETS    (CURRENT: {}   LAST: {}   VCID: {})".format(diff, vcdu.COUNTER, self.counter, vcdu.VCID))
                 else:
                     print("  DROPPED {} PACKETS".format(diff))
@@ -317,12 +307,12 @@ class Channel:
             # Close current TP_File
             lenok = self.cTPFile.finish(cppdu.PAYLOAD[:-2])
 
-            if self.verbose: self.cTPFile.print_info()
+            if self.config.verbose: self.cTPFile.print_info()
             if lenok:
-                if self.verbose: print("    LENGTH:     OK\n")
+                if self.config.verbose: print("    LENGTH:     OK\n")
                 
                 # Handle S_PDU (decryption)
-                spdu = CCSDS.S_PDU(self.cTPFile.PAYLOAD, self.keys)
+                spdu = CCSDS.S_PDU(self.cTPFile.PAYLOAD, self.config.keys)
 
                 # Handle xRIT file
                 self.handle_xRIT(spdu)
@@ -332,13 +322,13 @@ class Channel:
                 ac = len(self.cTPFile.PAYLOAD)
                 diff = ac - ex
 
-                if self.verbose: print("    LENGTH:     ERROR (EXPECTED: {}, ACTUAL: {}, DIFF: {})".format(ex, ac, diff))
+                if self.config.verbose: print("    LENGTH:     ERROR (EXPECTED: {}, ACTUAL: {}, DIFF: {})".format(ex, ac, diff))
                 print("  SKIPPING FILE (DROPPED PACKETS?)")
             
             # Clear finished TP_File
             self.cTPFile = None
         
-        if self.verbose:
+        if self.config.verbose:
             ac = len(self.cTPFile.PAYLOAD)
             ex = self.cTPFile.LENGTH
             p = round((ac/ex) * 100)
@@ -355,13 +345,14 @@ class Channel:
         xrit = CCSDS.xRIT(spdu.PLAINTEXT)
 
         # Save xRIT file if enabled
-        if self.outputXRIT:
-            xrit.save(self.outputPath)
+        if self.config.xrit:
+            xrit.save(self.config.output)
             xrit.print_info()
 
-        if self.outputImages:
+        if self.config.images:
             #TODO
             pass
+
 
     def notify(self, vcid):
         """
@@ -369,11 +360,11 @@ class Channel:
         """
 
         # No longer the active channel handler  
-        if vcid != self.VCID:
+        if vcid != self.config.VCID:
             # Channel has unfinished TP_File
             if self.cTPFile != None:
                 # Handle S_PDU (decryption)
-                spdu = CCSDS.S_PDU(self.cTPFile.PAYLOAD, self.keys)
+                spdu = CCSDS.S_PDU(self.cTPFile.PAYLOAD, self.config.keys)
 
                 # Handle xRIT file
                 self.handle_xRIT(spdu)